ghostty_terminal: Claude-status, sparade flik-sessioner och worktree-picker
Port av forkens ClaudeTabStatus/ClaudeTabIO: var 1,5 s samlas varje fliks
förgrunds-pids, i bakgrunden hittas claude via KERN_PROCARGS2, hook-filen
~/.claude/tab-state/<pid>.json läses, titeln tas ur transcriptens ai-title
och winmans bakgrundsjobb avgör blå lampa. done räknas som läst när fliken
har fokus, precis som i forken.

Flikarna sparas i ~/.config/ghostty/tab-sessions/<slug>.json, samma filer
som Ghostty-appen skrev, och återställs med claude --resume. En instans
startad med --user-data-dir håller sina sessioner i sin egen katalog.

Worktree-pickern under fliken listar <projekt>/worktrees/* med gren ur
.git, och cd:ar fliken med Ctrl-U och Return som tangenthändelser.
Blockerad-lampan öppnar forkens NSAlert-sheet för anteckningen.

+++ crates/ghostty_terminal/src/runtime.rs
@@ -442,7 +442,7 @@ unsafe extern "C" fn action_cb(
                 true
             }
             ffi::GHOSTTY_ACTION_GOTO_TAB => {
-                shared.send(SurfaceEvent::GotoTab(action.action.goto_tab as i32));
+                shared.send(SurfaceEvent::GotoTab(action.action.goto_tab));
                 true
             }
             ffi::GHOSTTY_ACTION_GOTO_SPLIT => {
